more docs, some refactoring
