+++ sqlite_storage/Cargo.toml
@@ -14,11 +14,7 @@ openmls_traits = { version = "0.3.0", path = "../traits" }
 thiserror = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 log = { version = "0.4" }
-serde_json = { version = "1.0", optional = true }
 rusqlite = { version = "0.32", features = ["bundled"] }
 
-[features]
-test-utils = ["serde_json", "openmls_traits/test-utils"] # Enable test utilites
-
 [dev-dependencies]
-openmls_sqlite_storage = { path = ".", features = ["test-utils"] }
+serde_json = { version = "1.0" }
