+++ sqlite_storage/src/lib.rs
@@ -5,12 +5,23 @@
 //!
 //! ## Usage
 //!
-//! TODO: Show how to combine with an rng and crypto provider in the context of
-//! OpenMLS. Also show how data (e.g. KeyPackages) can be retrieved by the
-//! application outside of OpenMLS.
+//! Generally, the [`SqliteStorageProvider`] can be used like any other storage
+//! provider. However, before first use, the tables need to be created. This can
+//! be done using the [`SqliteStorageProvider::create_tables`] method.
+//!
+//! ### Codec
+//!
+//! The [`SqliteStorageProvider`] can be instantiated with any codec that make
+//! use of the [`Serialize`] and [`DeserializeOwned`] traits of the `serde`
+//! crate. The codec is set by implementing [`Codec`] and passing the
+//! implementation as generic parameter to the [`SqliteStorageProvider`] upon
+//! creation.
 
 #[cfg(doc)]
 use openmls_traits::storage::StorageProvider;
+
+#[cfg(doc)]
+use serde::{de::DeserializeOwned, Serialize};
 
 mod codec;
 mod encryption_key_pairs;
@@ -24,7 +35,7 @@ mod signature_key_pairs;
 mod storage_provider;
 mod wrappers;
 
-pub use codec::{Codec, JsonCodec};
+pub use codec::Codec;
 pub use storage_provider::SqliteStorageProvider;
 
 trait Storable {
