+++ sqlite_storage/tests/proposals.rs
@@ -1,9 +1,24 @@
-use openmls_sqlite_storage::JsonCodec;
+use openmls_sqlite_storage::Codec;
 use openmls_traits::storage::{
     traits::{self},
     Entity, Key, StorageProvider, CURRENT_VERSION,
 };
 use serde::{Deserialize, Serialize};
+
+#[derive(Default)]
+pub struct JsonCodec;
+
+impl Codec for JsonCodec {
+    type Error = serde_json::Error;
+
+    fn to_vec<T: Serialize>(value: &T) -> Result<Vec<u8>, Self::Error> {
+        serde_json::to_vec(value)
+    }
+
+    fn from_slice<T: serde::de::DeserializeOwned>(slice: &[u8]) -> Result<T, Self::Error> {
+        serde_json::from_slice(slice)
+    }
+}
 
 // Test types
 #[derive(Serialize, Deserialize, PartialEq, Eq, Debug, Clone)]
